+++ src/hacks/precomputedNoiseTextures.ts
@@ -10,40 +10,36 @@ interface NoiseTextureData {
 	format: PIXI.FORMATS;
 }
 
-const FBM_TEXTURE_DATA: Record<string, NoiseTextureData> = {
-	'fbm(2)_5_10': {
-		path: `modules/${NAMESPACE}/dist/noise/fbm2-1_2.basis`,
+const FBM_TEXTURE_DATA = {
+	'fbm2+3': {
+		path: `modules/${NAMESPACE}/dist/noise/fbm2_3.avif`,
+		format: PIXI.FORMATS.RGB,
+	},
+	fbm4: {
+		path: `modules/${NAMESPACE}/dist/noise/fbm4.avif`,
 		format: PIXI.FORMATS.RG,
 	},
-	'fbm(3)_1_3_5': {
-		path: `modules/${NAMESPACE}/dist/noise/fbm3-1_3_5.avif`,
-		format: PIXI.FORMATS.RGB,
-	},
-	'fbm(4)_1_6_8': {
-		path: `modules/${NAMESPACE}/dist/noise/fbm4-1_6_8.avif`,
-		format: PIXI.FORMATS.RGB,
-	},
-	'fbmHQ(3)_1_10': {
-		path: `modules/${NAMESPACE}/dist/noise/fbmHQ3-1_10.basis`,
-		format: PIXI.FORMATS.RED,
+	fbmHQ3: {
+		path: `modules/${NAMESPACE}/dist/noise/fbmHQ3.basis`,
+		format: PIXI.FORMATS.RG,
 	},
 };
 
 // TODO remove me, this is just for debugging!
-// for (const [key, value] of Object.entries(FBM_TEXTURE_DATA)) {
-// 	FBM_TEXTURE_DATA[key].path = `${value.path}?v=${Math.random()}`;
-// }
+for (const [key, value] of Object.entries(FBM_TEXTURE_DATA)) {
+	FBM_TEXTURE_DATA[key].path = `${value.path}?v=${Math.random()}`;
+}
 
 const NOISE_TEXTURE_MAP: Record<string, NoiseTextureData> = {
-	ghost: FBM_TEXTURE_DATA['fbm(3)_1_3_5'],
-	fairy: FBM_TEXTURE_DATA['fbm(3)_1_3_5'],
-	witchwave: FBM_TEXTURE_DATA['fbm(4)_1_6_8'],
-	fog: FBM_TEXTURE_DATA['fbm(4)_1_6_8'],
-	vortex: FBM_TEXTURE_DATA['fbm(4)_1_6_8'],
-	smokepatch: FBM_TEXTURE_DATA['fbmHQ(3)_1_10'],
-	hole: FBM_TEXTURE_DATA['fbmHQ(3)_1_10'],
-	dome: FBM_TEXTURE_DATA['fbm(2)_5_10'],
-	roiling: FBM_TEXTURE_DATA['fbm(3)_1_3_5'],
+	ghost: FBM_TEXTURE_DATA['fbm2+3'],
+	fairy: FBM_TEXTURE_DATA['fbm2+3'],
+	witchwave: FBM_TEXTURE_DATA.fbm4,
+	fog: FBM_TEXTURE_DATA.fbm4,
+	vortex: FBM_TEXTURE_DATA.fbm4,
+	smokepatch: FBM_TEXTURE_DATA.fbmHQ3,
+	hole: FBM_TEXTURE_DATA.fbmHQ3,
+	dome: FBM_TEXTURE_DATA['fbm2+3'],
+	roiling: FBM_TEXTURE_DATA['fbm2+3'],
 };
 
 const NOISE_TEXTURE_URLS = Array.from(new Set(Object.values(NOISE_TEXTURE_MAP)));
@@ -80,7 +76,7 @@ function AdaptiveLightingShader__updateDarknessUniforms(this: BaseLightSource<an
 	u.fbmTexture = FOUNDRY_API.getTexture(noiseTextureData.path);
 }
 
-function optimizedFBM() {
+function optimizedFBMOld() {
 	return `
 
 uniform sampler2D fbmTexture;
@@ -100,91 +96,123 @@ float fbmLQ(in float factor, in float channel, in vec2 uv, in float ignored_smoo
 `;
 }
 
+interface FBMData {
+	scale: number;
+	channel: 'r' | 'g' | 'b' | 'a';
+	boost?: number;
+	name?: string;
+}
+function optimizedFBM(...fbmData: FBMData[]) {
+	const functions = fbmData
+		.map(({ channel, scale, name, boost = 1 }) => {
+			const boostCalculation = boost === 1 ? '' : ` * ${boost}`;
+			const scaleRec = 1 / 5 / scale;
+			return `
+float fbm${name}(in vec2 uv) {
+  vec4 color = texture2D(fbmTexture, uv * ${scaleRec.toPrecision(4)});
+  return color.${channel}${boostCalculation};
+}
+float fbm${name}(in vec2 uv, in float ignoredSmoothness) {
+  return fbm${name}(uv);
+}
+		`;
+		})
+		.join('\n');
+	return `uniform sampler2D fbmTexture;\n${functions}`;
+}
+
 function patchShaderFBM() {
 	const getShaderByName = (name: string) => {
 		return FOUNDRY_API.generation < 13 ? eval(name) : foundry.canvas.rendering.shaders[name];
 	};
-	const fbm21 = getShaderByName('GhostLightColorationShader').FBM(2, 1);
-	const fbm31 = getShaderByName('GhostLightColorationShader').FBM(3, 1);
-	const fbm41 = getShaderByName('GhostLightColorationShader').FBM(4, 1);
+	const fbm2 = getShaderByName('GhostLightColorationShader').FBM(2, 1);
+	const fbm3 = getShaderByName('GhostLightColorationShader').FBM(3, 1);
+	const fbm4 = getShaderByName('GhostLightColorationShader').FBM(4, 1);
 	const fbmHQ3 = getShaderByName('GhostLightColorationShader').FBMHQ(3);
+
+	const optimizedFBM2 = optimizedFBM({ name: '2', channel: 'r', scale: 1.6 });
+	const optimizedFBM3 = optimizedFBM({ name: '3_1', channel: 'g', scale: 1 }, { name: '3_5', channel: 'b', scale: 5 });
+	const optimizedFBM4 = optimizedFBM({ name: '4_1', channel: 'r', scale: 1 }, { name: '4_8', channel: 'g', scale: 8 });
+	const optimizedFBMHQ3 = optimizedFBM(
+		{ name: 'HQ3_1', channel: 'r', scale: 1 },
+		{ name: 'HQ3_10', channel: 'g', scale: 5 },
+	);
 
 	const shaders = [
 		{
 			shader: 'GhostLightIlluminationShader',
 			replacements: [
-				[fbm31, fbm31 + optimizedFBM()],
+				[fbm3, fbm3 + optimizedFBM3],
 				[
 					// wrap-line
 					'float distortion1 = fbm(vec2(',
-					'float distortion1 = fbmLQ(1.0, 1.0, vec2(',
+					'float distortion1 = fbm3_1(vec2(',
 				],
 				[
 					// wrap-line
 					'fbm(vUvs * 5.0 - time * 0.50),',
-					'fbmLQ(5.0, 3.0, vUvs * 5.0 - time * 0.50),',
+					'fbm3_5(vUvs * 5.0 - time * 0.50),',
 				],
 				[
 					// wrap-line
-
 					'fbm((-vUvs - vec2(0.01)) * 5.0 + time * INVTHREE)));',
-					'fbmLQ(5.0, 3.0, (-vUvs - vec2(0.01)) * 5.0 + time * INVTHREE)));',
+					'fbm3_5((-vUvs - vec2(0.01)) * 5.0 + time * INVTHREE)));',
 				],
 				[
 					// wrap-line
 					'float distortion2 = fbm(vec2(',
-					'float distortion2 = fbmLQ(1.0, 1.0, vec2(',
+					'float distortion2 = fbm3_1(vec2(',
 				],
 				[
 					// wrap-line
 					'fbm(-vUvs * 5.0 - time * 0.50),',
-					'fbmLQ(5.0, 3.0, -vUvs * 5.0 - time * 0.50),',
+					'fbm3_5(-vUvs * 5.0 - time * 0.50),',
 				],
 				[
 					// wrap-line
-
 					'fbm((-vUvs + vec2(0.01)) * 5.0 + time * INVTHREE)));',
-					'fbmLQ(5.0, 3.0, (-vUvs + vec2(0.01)) * 5.0 + time * INVTHREE)));',
+					'fbm3_5((-vUvs + vec2(0.01)) * 5.0 + time * INVTHREE)));',
 				],
 			],
 		},
 		{
 			shader: 'GhostLightColorationShader',
 			replacements: [
-				[fbm31, fbm31 + optimizedFBM()],
+				[fbm3, fbm3 + optimizedFBM3],
 				[
 					// wrap-line
 					'float distortion1 = fbm(vec2(',
-					'float distortion1 = fbmLQ(1.0, 1.0, vec2(',
+					'float distortion1 = fbm3_1(vec2(',
 				],
 				[
 					// wrap-line
 					'fbm(vUvs * 3.0 + time * 0.50),',
-					'fbmLQ(3.0, 2.0, vUvs * 3.0 + time * 0.50),',
+					'fbm3_5(vUvs * 3.0 + time * 0.50),',
 				],
 				[
+					// wrap-line
 					'fbm((-vUvs + vec2(1.)) * 5.0 + time * INVTHREE)));',
-					'fbmLQ(5.0, 3.0, (-vUvs + vec2(1.)) * 5.0 + time * INVTHREE)));',
+					'fbm3_5((-vUvs + vec2(1.)) * 5.0 + time * INVTHREE)));',
 				],
 				[
 					// wrap-line
 					'float distortion2 = fbm(vec2(',
-					'float distortion2 = fbmLQ(1.0, 1.0, vec2(',
+					'float distortion2 = fbm3_1(vec2(',
 				],
 				[
 					// wrap-line
 					'fbm(-vUvs * 3.0 + time * 0.50),',
-					'fbmLQ(3.0, 2.0, -vUvs * 3.0 + time * 0.50),',
+					'fbm3_5(-vUvs * 3.0 + time * 0.50),',
 				],
 				[
 					// wrap-line
 					'fbm((-vUvs + vec2(1.)) * 5.0 - time * INVTHREE)));',
-					'fbmLQ(5.0, 3.0,(-vUvs + vec2(1.)) * 5.0 - time * INVTHREE)));',
+					'fbm3_5((-vUvs + vec2(1.)) * 5.0 - time * INVTHREE)));',
 				],
 				[
 					// wrap-line
 					'uv *= fbm(vec2(time + distortion1, time + distortion2));',
-					'uv *= fbmLQ(1.0, 1.0, vec2(time + distortion1, time + distortion2));',
+					'uv *= fbm3_1(vec2(time + distortion1, time + distortion2));',
 				],
 			],
 		},
@@ -192,80 +220,74 @@ function patchShaderFBM() {
 		{
 			shader: 'FairyLightIlluminationShader',
 			replacements: [
-				[fbm31, fbm31 + optimizedFBM()],
+				[fbm3, fbm3 + optimizedFBM3],
 				// distortion 1
 				[
 					// wrap-line
 					'float distortion1 = fbm(vec2(',
-					'float distortion1 = fbmLQ(1.0, 1.0, vec2(',
+					'float distortion1 = fbm3_1(vec2(',
 				],
 				[
 					// wrap-line
 					'fbm(vUvs * 3.0 - time * 0.50), ',
-					'fbmLQ(3.0, 2.0, vUvs * 3.0 - time * 0.50),',
+					'fbm3_5(vUvs * 3.0 - time * 0.50),',
 				],
-				[
-					'fbm((-vUvs + vec2(1.)) * 5.0 + time * INVTHREE)));',
-					'fbmLQ(5.0, 3.0, (-vUvs + vec2(1.)) * 5.0 + time * INVTHREE)));',
-				],
+				['fbm((-vUvs + vec2(1.)) * 5.0 + time * INVTHREE)));', 'fbm3_5((-vUvs + vec2(1.)) * 5.0 + time * INVTHREE)));'],
 				// distortion 2
 				[
 					// wrap-line
 					'float distortion2 = fbm(vec2(',
-					'float distortion2 = fbmLQ(1.0, 1.0, vec2(',
+					'float distortion2 = fbm3_1(vec2(',
 				],
 				[
 					// wrap-line
 					'fbm(-vUvs * 3.0 - time * 0.50),',
-					'fbmLQ(3.0, 2.0, -vUvs * 3.0 - time * 0.50),',
+					'fbm3_5(-vUvs * 3.0 - time * 0.50),',
 				],
 				[
 					// wrap-line
 					'fbm((-vUvs + vec2(1.)) * 5.0 - time * INVTHREE)));',
-					'fbmLQ(5.0, 3.0, (-vUvs + vec2(1.)) * 5.0 - time * INVTHREE)));',
+					'fbm3_5((-vUvs + vec2(1.)) * 5.0 - time * INVTHREE)));',
 				],
 			],
 		},
 		{
 			shader: 'FairyLightColorationShader',
 			replacements: [
-				[fbm31, fbm31 + optimizedFBM()],
+				[fbm3, fbm3 + optimizedFBM3],
 				// distortion 1
 				[
 					// wrap-line
 					'float distortion1 = fbm(vec2(',
-					'float distortion1 = fbmLQ(1.0, 1.0, vec2(',
+					'float distortion1 = fbm3_1(vec2(',
 				],
 				[
 					// wrap-line
 					'fbm(vUvs * 3.0 + time * 0.50),',
-					'fbmLQ(3.0, 2.0, vUvs * 3.0 + time * 0.50),',
+					'fbm3_5(vUvs * 3.0 + time * 0.50),',
 				],
-				[
-					'fbm((-vUvs + vec2(1.)) * 5.0 + time * INVTHREE)));',
-					'fbmLQ(5.0, 3.0, (-vUvs + vec2(1.)) * 5.0 + time * INVTHREE)));',
-				],
+				['fbm((-vUvs + vec2(1.)) * 5.0 + time * INVTHREE)));', 'fbm3_5((-vUvs + vec2(1.)) * 5.0 + time * INVTHREE)));'],
 				// distortion 2
 				[
 					// wrap-line
 					'float distortion2 = fbm(vec2(',
-					'float distortion2 = fbmLQ(1.0, 1.0, vec2(',
+					'float distortion2 = fbm3_1(vec2(',
 				],
 				[
 					// wrap-line
 					'fbm(-vUvs * 3.0 + time * 0.50),',
-					'fbmLQ(3.0, 2.0, -vUvs * 3.0 + time * 0.50),',
+					'fbm3_5(-vUvs * 3.0 + time * 0.50),',
 				],
 				[
 					// wrap-line
 					'fbm((-vUvs + vec2(1.)) * 5.0 - time * INVTHREE)));',
-					'fbmLQ(5.0, 3.0, (-vUvs + vec2(1.)) * 5.0 - time * INVTHREE)));',
+					'fbm3_5((-vUvs + vec2(1.)) * 5.0 - time * INVTHREE)));',
 				],
 				// final FBM
 				[
 					// wrap-line
 					'uv *= fbm(vec2(time + distortion1, time + distortion2));',
-					'uv *= fbmLQ(1.0, 1.0, vec2(time + distortion1, time + distortion2));',
+					'uv *= fbm3_1(vec2(time + distortion1, time + distortion2));',
 				],
 			],
 		},
@@ -273,22 +295,22 @@ function patchShaderFBM() {
 		{
 			shader: 'BewitchingWaveIlluminationShader',
 			replacements: [
-				[fbm41, fbm41 + optimizedFBM()],
+				[fbm4, fbm4 + optimizedFBM4],
 				[
 					// wrap-line
 					'float motion = fbm(uv + time * 0.25);',
-					'float motion = fbmLQ(1.0, 1.0, uv + time * 0.25);',
+					'float motion = fbm4_1(uv + time * 0.25);',
 				],
 			],
 		},
 		{
 			shader: 'BewitchingWaveColorationShader',
 			replacements: [
-				[fbm41, fbm41 + optimizedFBM()],
+				[fbm4, fbm4 + optimizedFBM4],
 				[
 					// wrap-line
 					'float motion = fbm(uv + time * 0.25);',
-					'float motion = fbmLQ(1.0, 1.0, uv + time * 0.25);',
+					'float motion = fbm4_1(uv + time * 0.25);',
 				],
 			],
 		},
@@ -296,26 +318,26 @@ function patchShaderFBM() {
 		{
 			shader: 'FogColorationShader',
 			replacements: [
-				[fbm41, fbm41 + optimizedFBM()],
+				[fbm4, fbm4 + optimizedFBM4],
 				[
 					// wrap-line
 					'float q = fbm(p - time * 0.1);',
-					'float q = fbmLQ(8.0, 3.0, p - time * 0.1);',
+					'float q = fbm4_8(p - time * 0.1);',
 				],
 				[
 					// wrap-line
 					'vec2 r = vec2(fbm(p + q - time * 0.5 - p.x - p.y),',
-					'vec2 r = vec2(fbmLQ(1.0, 1.0, p + q - time * 0.5 - p.x - p.y),',
+					'vec2 r = vec2(fbm4_8(p + q - time * 0.5 - p.x - p.y),',
 				],
 				[
 					// wrap-line
 					'fbm(p + q - time * 0.3));',
-					'fbmLQ(8.0, 3.0, p + q - time * 0.3));',
+					'fbm4_8(p + q - time * 0.3));',
 				],
 				[
 					// wrap-line
 					'fbm(p + r)) + mix(c3, c4, r.x)',
-					'fbmLQ(8.0, 3.0, p + r)) + mix(c3, c4, r.x)',
+					'fbm4_8(p + r)) + mix(c3, c4, r.x)',
 				],
 			],
 		},
@@ -323,47 +345,47 @@ function patchShaderFBM() {
 		{
 			shader: 'VortexIlluminationShader',
 			replacements: [
-				[fbm41, fbm41 + optimizedFBM()],
+				[fbm4, fbm4 + optimizedFBM4],
 				[
 					// wrap-line
 					'float q = fbm(p + time);',
-					'float q = fbmLQ(6.0, 1.0, p + time);',
+					'float q = fbm4_8(p + time);',
 				],
 				[
 					// wrap-line
 					'vec2 r = vec2(fbm(p + q + time * 0.9 - p.x - p.y), fbm(p + q + time * 0.6));',
-					'vec2 r = vec2(fbmLQ(1.0, 1.0, p + q + time * 0.9 - p.x - p.y), fbmLQ(6.0, 2.0, p + q + time * 0.6));',
+					'vec2 r = vec2(fbm4_8(p + q + time * 0.9 - p.x - p.y), fbm4_8(p + q + time * 0.6));',
 				],
 				[
 					// wrap-line
 					'return mix(c1, c2, fbm(p + r)) + mix(c3, c4, r.x) - mix(c5, c6, r.y);',
-					'return mix(c1, c2, fbmLQ(6.0, 2.0, p + r)) + mix(c3, c4, r.x) - mix(c5, c6, r.y);',
+					'return mix(c1, c2, fbm4_8(p + r)) + mix(c3, c4, r.x) - mix(c5, c6, r.y);',
 				],
 			],
 		},
 		{
 			shader: 'VortexColorationShader',
 			replacements: [
-				[fbm41, fbm41 + optimizedFBM()],
+				[fbm4, fbm4 + optimizedFBM4],
 				[
 					// wrap-line
 					'float q = fbm(p + time);',
-					'float q = fbmLQ(6.0, 2.0, p + time);',
+					'float q = fbm4_8(p + time);',
 				],
 				[
 					// wrap-line
 					'vec2 r = vec2(fbm(p + q + time * 0.9 - p.x - p.y), ',
-					'vec2 r = vec2(fbmLQ(1.0, 1.0, p + q + time * 0.9 - p.x - p.y), ',
+					'vec2 r = vec2(fbm4_1(p + q + time * 0.9 - p.x - p.y), ',
 				],
 				[
 					// wrap-line
 					'fbm(p + q + time * 0.6));',
-					'fbmLQ(6.0, 2.0, p + q + time * 0.6));',
+					'fbm4_8(p + q + time * 0.6));',
 				],
 				[
 					// wrap-line
 					'fbm(p + r)) + mix(c3, c4, r.x) ',
-					'fbmLQ(6.0, 2.0, p + r)) + mix(c3, c4, r.x) ',
+					'fbm4_8(p + r)) + mix(c3, c4, r.x) ',
 				],
 			],
 		},
@@ -371,32 +393,32 @@ function patchShaderFBM() {
 		{
 			shader: 'SmokePatchIlluminationShader',
 			replacements: [
-				[fbmHQ3, fbmHQ3 + optimizedFBM()],
+				[fbmHQ3, fbmHQ3 + optimizedFBMHQ3],
 				[
 					// wrap-line
 					'max(fbm(uv + t, 1.0),',
-					'max(fbmLQ(1.0, 1.0, uv + t, 1.0),',
+					'max(fbmHQ3_1(uv + t, 1.0),',
 				],
 				[
 					// wrap-line
 					'fbm(uv - t, 1.0)),',
-					'fbmLQ(1.0, 1.0, uv - t, 1.0)),',
+					'fbmHQ3_1(uv - t, 1.0)),',
 				],
 			],
 		},
 		{
 			shader: 'SmokePatchColorationShader',
 			replacements: [
-				[fbmHQ3, fbmHQ3 + optimizedFBM()],
+				[fbmHQ3, fbmHQ3 + optimizedFBMHQ3],
 				[
 					// wrap-line
 					'max(fbm(uv + t, 1.0),',
-					'max(fbmLQ(10.0, 1.0, uv + t, 1.0),',
+					'max(fbmHQ3_1(uv + t, 1.0),',
 				],
 				[
 					// wrap-line
 					'fbm(uv - t, 1.0)),',
-					'fbmLQ(10.0, 1.0, uv - t, 1.0)),',
+					'fbmHQ3_1(uv - t, 1.0)),',
 				],
 			],
 		},
@@ -404,11 +426,11 @@ function patchShaderFBM() {
 		{
 			shader: 'BlackHoleDarknessShader',
 			replacements: [
-				[fbmHQ3, fbmHQ3 + optimizedFBM()],
+				[fbmHQ3, fbmHQ3 + optimizedFBMHQ3],
 				[
 					// wrap-line
 					'float beams = fract(angle + sin(dist * 30.0 * (intensity * 0.2) - time + fbm(uv * 10.0 + time * 0.25, 1.0) * dad));',
-					'float beams = fract(angle + sin(dist * 30.0 * (intensity * 0.2) - time + fbmLQ(10.0, 1.0, uv * 10.0 + time * 0.25, 1.0) * dad));',
+					'float beams = fract(angle + sin(dist * 30.0 * (intensity * 0.2) - time + fbmHQ3_10(uv * 10.0 + time * 0.25, 1.0) * dad));',
 				],
 			],
 		},
@@ -416,21 +438,21 @@ function patchShaderFBM() {
 		{
 			shader: 'LightDomeColorationShader',
 			replacements: [
-				[fbm21, fbm21 + optimizedFBM()],
+				[fbm2, fbm2 + optimizedFBM2],
 				[
 					// wrap-line
 					'float q = 2.0 * fbm(p + time * 0.2);',
-					'float q = 2.0 * fbmLQ(1.0, 1.0, p + time * 0.2);',
+					'float q = 2.0 * fbm2(p + time * 0.2);',
 				],
 				[
 					// wrap-line
 					'vec2 r = vec2(fbm(p + q + ( time  ) - p.x - p.y), fbm(p * 2.0 + ( time )));',
-					'vec2 r = vec2(fbmLQ(1.0, 1.0, p + q + ( time  ) - p.x - p.y), fbmLQ(2.0, 2.0, p * 2.0 + ( time )));',
+					'vec2 r = vec2(fbm2(p + q + ( time  ) - p.x - p.y), fbm2(p * 2.0 + ( time )));',
 				],
 				[
 					// wrap-line
 					'return clamp( mix( c1, c2, abs(fbm(p + r)) ) + mix( c3, c4, abs(r.x * r.x * r.x) ) - mix(',
-					'return clamp( mix( c1, c2, fbmLQ(1.0, 1.0, p + r) ) + mix( c3, c4, abs(r.x * r.x * r.x) ) - mix(',
+					'return clamp( mix( c1, c2, fbm2(p + r) ) + mix( c3, c4, abs(r.x * r.x * r.x) ) - mix(',
 				],
 			],
 		},
@@ -438,37 +460,37 @@ function patchShaderFBM() {
 		{
 			shader: 'RoilingDarknessShader',
 			replacements: [
-				[fbm31, fbm31 + optimizedFBM()],
+				[fbm3, fbm3 + optimizedFBM3],
 				[
 					// wrap-line
 					`float distortion1 = fbm( vec2(`,
-					`float distortion1 = fbmLQ(1.0, 1.0, vec2(`,
+					`float distortion1 = fbm3_1(vec2(`,
 				],
 				[
 					// wrap-line
 					`fbm( vUvs * 2.5 + time * 0.5),`,
-					`fbmLQ(3.0, 2.0, vUvs * 2.5 + time * 0.5),`,
+					`fbm3_5(vUvs * 2.5 + time * 0.5),`,
 				],
 				[
 					// wrap-line
 					`fbm( (-vUvs - vec2(0.01)) * 5.0 + time * INVTHREE)));`,
-					`fbmLQ(5.0, 3.0, (-vUvs - vec2(0.01)) * 5.0 + time * INVTHREE)));`,
+					`fbm3_5((-vUvs - vec2(0.01)) * 5.0 + time * INVTHREE)));`,
 				],
 
 				[
 					// wrap-line
 					`float distortion2 = fbm( vec2(`,
-					`float distortion2 = fbmLQ(1.0, 1.0, vec2(`,
+					`float distortion2 = fbm3_1(vec2(`,
 				],
 				[
 					// wrap-line
 					`fbm( -vUvs * 5.0 + time * 0.5),`,
-					`fbmLQ(5.0, 3.0, -vUvs * 5.0 + time * 0.5),`,
+					`fbm3_5(-vUvs * 5.0 + time * 0.5),`,
 				],
 				[
 					// wrap-line
 					`fbm( (vUvs + vec2(0.01)) * 2.5 + time * INVTHREE)));`,
-					`fbmLQ(3.0, 2.0, (vUvs + vec2(0.01)) * 2.5 + time * INVTHREE)));`,
+					`fbm3_5((vUvs + vec2(0.01)) * 2.5 + time * INVTHREE)));`,
 				],
 			],
 		},
@@ -514,6 +536,8 @@ async function enablePrecomputedNoiseTextures() {
 			texture.baseTexture.wrapMode = PIXI.WRAP_MODES.REPEAT;
 			texture.baseTexture.format = format;
 			texture.baseTexture.scaleMode = PIXI.SCALE_MODES.LINEAR;
+			texture.baseTexture.alphaMode = PIXI.ALPHA_MODES.NPM;
+			texture.baseTexture.mipmap = PIXI.MIPMAP_MODES.OFF;
 		}),
 	);
 }